+++ src/preload.js
@@ -14,6 +14,7 @@ const { contextBridge, ipcRenderer } = require('electron');
 contextBridge.exposeInMainWorld('claudeAPI', {
   getData:            ()                  => ipcRenderer.invoke('get-data'),
   refreshHookDeps:    ()                  => ipcRenderer.invoke('hooks:refresh-deps'),
+  checkHookTool:      (tool)              => ipcRenderer.invoke('hooks:check-tool', { tool }),
   pluginAction:       (action, pluginId)  => ipcRenderer.invoke('plugin-action',      { action, pluginId }),
   marketplaceAction:  (action, name, src) => ipcRenderer.invoke('marketplace-action', { action, name, source: src }),
   getMarketplaceDetail: (name)            => ipcRenderer.invoke('get-marketplace-detail', name),